Se implementa la BR-PEND-001 para validar que los pendientes son asocaidos al concpeto especial después de ser grabados.

+++ SemantikosKernelEJB/src/cl/minsal/semantikos/kernel/components/CategoryManagerImpl.java
@@ -89,7 +89,7 @@ public class CategoryManagerImpl implements CategoryManager {
     }
 
     @Override
-    public Category getCategoryById(int id) {
+    public Category getCategoryById(long id) {
         return categoryDAO.getCategoryById(id);
     }
